feat: refactor useStudentStatistics hook with type definitions - Extract type definitions to separate types.ts file - Improve hook structure and maintainability - Add proper TypeScript interfaces for statistics data
